Fix bad bug in string repr for `double` being `"float32"`.  Mostly only impacted emissions of data for humans like `nio meta`, confusing the poor human reader.

+++ nio.nim
@@ -14,7 +14,7 @@ from memfiles as mf import nil
 type #*** BASIC TYPE SETUP  #NOTE: gcc __float128 CPU-portable but slow
   IOKind* = enum cIk = "int8" , CIk = "uint8" , sIk = "int16", SIk = "uint16",
                  iIk = "int32", IIk = "uint32", lIk = "int64", LIk = "uint64",
-                 fIk = "float32", dIk = "float32", gIk = "float80"
+                 fIk = "float32", dIk = "float64", gIk = "float80"
 
   IOCol* = object       ## column metadata
     iok*:   IOKind      ## number kind
